pyproject.toml
[tool.poetry]
name = "monitorweb"
version = "0.1.0"
description = ""
authors = ["mikheyev9 <mikheyev9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
vt-py = "^0.18.3"
python-dotenv = "^1.0.1"
jupyter = "^1.0.0"
nest-asyncio = "^1.6.0"
sqlalchemy = "^2.0.32"
aiosqlite = "^0.20.0"
aiogram = "^3.10.0"
colorama = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
